requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
langgraph==0.0.20
faiss-cpu==1.7.4
mysql-connector-python==8.2.0
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.6.1
numpy==1.24.3
sentence-transformers==2.2.2